fix(stylelint): bound fallback analysis
